libsharedobjs: Create config option for sharedlib.
This commit introduces the config option to enable the creation of the
shared objects library for a particular board.  To use, simply define it
in board.h.

   #define CONFIG_SHAREDLIB

In order to actually add objects to the shared library, one must utilize
the SHAREDLIB() macro around the definition of the object.  All of the
objects in the shared library will be placed into the '.roshared'
section.  The SHAREDLIB() macro simply adds this attribute and prevents
the RW image from compiling them in.

For example,

   # In foo/bar_shared.c
   SHAREDLIB(const uint8_t shared_var = 0x43);

   # In include/bar_shared.h
   extern const uint8_t shared_var;

   # In include/libsharedobjs.h
   /* List of shared object module includes */
   #include "bar_shared.h"

BUG=none
BRANCH=none
TEST=make -j buildall tests
TEST=Enabled config option and saw the sharedlib being created.
TEST=Enabled config option on unsupported board and saw build fail.
CQ-DEPEND=CL:274079

Change-Id: I2abeb1be248ab161fa81c897d2f5793f5a599456
Signed-off-by: Aseda Aboagye <aaboagye@google.com>