Update patch set 1
Patch Set 1: Verified+1

Thanks for taking a look at this CL. There's probably a ton wrong here and a lot of room for improvement, but I just wanted to start introducing my changes instead of waiting to land a giant CL later on.

However, with this patch I was able to test a proof-of-concept on samus by sharing the 8042 scancode table with both RO and RW. This was a savings of 504 bytes.

Once this build infrastructure patch lands, I'll start to submit the next stages for review.

Patch-set: 1
Label: Verified=+1