Update patch set 1
Patch Set 1:

One benefit to moving the shared code to separate modules is that it'll make it more explicit what the dependencies are on data.

Otherwise it'll be easy for someone to modify a module such that the new non-shared functions don't work with the previous version of the shared functions.

Patch-set: 1