Update patch set 8
Patch Set 8:

We checked this version of the driver on glados. Seems to charge. Not sure how to handle status state across a sysjump w/o adding all that code to save/restore stuff. The current is pulled back in from the register state so the code assumes the charge state machine will catch any weirdness around the sysjump boundary.

Patch-set: 8