Update patch set 1
Patch Set 1:

No, it's not the same as CONFIG_FLASH_MAPPED.

That means flash is memory-mapped, but there may be a significant performance impact to executing code from mapped flash vs. SRAM - and it also may not work at all in the case where we're sharing a SPI flash between EC and AP.

So we need a separate #define to indicate that we want to execute from flash (particularly in the EC-RW code).

Patch-set: 1