Update patch set 4
Patch Set 4:

This is fine now, except our geometry CONFIGs have special meaning for npcx, and we're probably breaking npcx now that we fixed things for mec1322.

We can either put this change on hold until the fix is done, or modify vboot_hash.c so that it works using the old constants for npcx. If you decide on the latter, it's probably best to add a single ifdef at the top of the file only, along with a "TODO(crosbug.com/p/23796): Unify meaning of geometry constants...".

Patch-set: 4