glados: Inhibit AP power-on until charge current limit is set
It's sometimes desirable to boot without a battery, but we may brown out
if we don't have sufficient current. Inhibit AP power-on through the
BATLOW pin, even if the system is unprotected, until our charger and
current limit are initialized.

BUG=chrome-os-partner:41258
TEST=Manual on glados. Remove battery and attach Zinger. Verify EC
powers on and AP doesn't boot. Run `powerbtn`, verify that AP boots.
Remove all power and attach battery, verify that EC powers on and AP
boots.
BRANCH=None

Change-Id: Ifc3d16f8288a035854e9fd05812ce6de33170d6a
Signed-off-by: Shawn Nematbakhsh <shawnn@chromium.org>