Update patch set 6
Patch Set 6:

so the problem now is, since spi_enable()/disable is launched more often we are running into watchdog when flashing spi with host cmd.

As per spec, flash write takes max 3ms for 256 block. Since we are programming 1024 blocks, this tranlsates to >3s, and watchdog is set to 1.6s.

Is it ok to add watchdog reload in this patch or shall we start a new issue for this?

Patch-set: 6