Update patch set 6
Patch Set 6:

> It's all copied from lm4. Note the comments in lm4:
 > 
 > /*
 > * Reload the watchdog timer, so that erasing many flash pages
 > * doesn't cause a watchdog reset.  May not need this now that
 > * we're using msleep() below.
 > */
 > watchdog_reload();
 > 
 > We should be using *sleep() rather than *delay() except when we
 > must have high precision in our delay time.

the uber way to do it, would be to post a timer and handle result on expiration. It is probably a good idea to have a generic way to do it, as we have bunch of polling.

the way current code works is that I believe we just do spi_wait() which just polls status registers until it says not-busy.

Anyway, shall we have a new ticket for this, or should we stuff this code in this CL?

Patch-set: 6