Update patch set 6
Patch Set 6:

> it occured to me that msleep() will trigger rescheduling and that
 > will make watchdog happy. So no need to reload watchdog

I discovered an interesting jewel.

Looks like regular watchdog_reload() is sitting on HOOK_TICK. However it does not get a chance to tick (for whatever reason) during flash write scenario. With hook debug on I am seeing this:

102.167000 Hook at interval 250000 us delayed by 1357099 us

so it looks like it is already delayed by 1.3 seconds. 

In this case *sleep() will not help. In fact, spi_flash_wait() already does usleep, but it is not triggering watchdog reload, and now it is a bit more clear why.

(just for a record watchdog_reload() on each page program does solve the issue, although it is bad solution)

Patch-set: 6