Update patch set 6
Patch Set 6:

The WDT is normally reloaded every 250ms from the hooks task, which is the lowest priority task. It will fail to be reloaded if the hooks tasks starves, and that will happen if there are other tasks that are always ready. udelay() is a busy wait. usleep() tells the scheduler that the task isn't ready for the next X us so other ready tasks (such as hooks) can run.

But we need to pay close attention to the context switch latency. If our usleep() delay is less than our context switch latency then usleep() is no better (and probably worse) than udelay().

Patch-set: 6