Update patch set 6
Patch Set 6:

> The WDT is normally reloaded every 250ms from the hooks task, which
 > is the lowest priority task. It will fail to be reloaded if the
 > hooks tasks starves, and that will happen if there are other tasks
 > that are always ready. udelay() is a busy wait. usleep() tells the
 > scheduler that the task isn't ready for the next X us so other
 > ready tasks (such as hooks) can run.
 > 
 > But we need to pay close attention to the context switch latency.
 > If our usleep() delay is less than our context switch latency then
 > usleep() is no better (and probably worse) than udelay().

I think it is plausible that hook timer does not have a chance to fire. mec1322 is running on 48mhz, which is 21ns per clock cycle. Now then, as part of spi_flash_wait() we usleep() for 100 usec. So we are getting roughly 4500 cycles. Given some other latencies on the way we may not meet the deadline, right?

So in my opinion what we need to do is to add msleep() for 1.5 msec, which is the "typical" page programming time. 1.5ms. This will likely let the hook to fire.

I can try this out (although I can not be 100% confident of result because reproducibility is not perfect in the first place)

Patch-set: 6
Label: Code-Review=0