Cr50: Put SPI slave into loopback mode
This just causes all incoming bytes on the SPI slave bus to be
sent right back out again, so that we can test the I/O.

BUG=chrome-os-partner:34995
BRANCH=none
TEST=make buildall

I also tested this using Vadim's spiraw.py script, and a one-off
compiled libmpsse program. There seems to be some significant lag
between what we send and what we receive - about 14 bytes or so -
but I'm not 100% certain that it's all due to the slow response
time of the Cr50. There may be driver bugs, or I may be using the
FTDI and MPSSE libraries incorrectly.

Signed-off-by: Bill Richardson <wfrichar@chromium.org>
Change-Id: Id6a8a025f61cc26e5c1d1f7bb3bc9826ca56a005