Update patch set 1
Patch Set 1:

No, this doesn't touch the blob task. It's just an immediate echo, which is what you said you wanted for testing.

The trunks_client test expects stall cycles, which we can't provide, so I'll have to change the protocol in order to communicate with the blob. That'll be another CL.

Patch-set: 1