Update patch set 1
Patch Set 1:

I will change the naming but moving it to upper layer.. Not for this patch. There is enough testing done for this patch and I would not want to theoretically predict any behavior and change the timing. I can move it later and we have BKC testing on a regular basis. so we will upload a tested code.

Patch-set: 1
Label: Code-Review=0