pd: Change which DP pin config DP source uses for sink capabilities.
DP Alternate mode specification defines two bytes in the DP
capabilities payload with respect to pin configs.  One that governs
the DFP and the other the UFP.  The one that the DP source needs to
examine depends upon whether the UFP is a receptacle or a plug.

Previously we just examined the DFP config which was correct for all
UFP plugs but not UFP receptacles.  This change fixes that by
examining bit <6> of capability message to determing plug vs
receptacle.

Signed-off-by: Todd Broch <tbroch@chromium.org>

BRANCH=samus
BUG=chromium:505652
TEST=manual, hoho still works.

Change-Id: Ib9e3a808828ff6b8c062baf0fc3c23b3da33c2c3