Add EC_CMD_BLOB host command to pass opaque data around
This provides a way to experiment with or support random
functionality not specifically related to the normal EC
operation. With this, we can use a standard host command to pass
opaque data blobs between the AP and EC without specifying a
particular meaning to them.

BUG=chrome-os-partner:40969
BRANCH=none
TEST=make buildall

Nothing in the normal EC usage sends these messages. I have a
test/example program coming soon.

Change-Id: I77fe434cbd9151e4c2c22bc36d1ed6063d3bf940
Signed-off-by: Bill Richardson <wfrichar@chromium.org>