Cr50: Enable host commands over SPI bus
This enables the feature that lets the Cr50 receive host commands
via the SPI (slave) interface.

BUG=chrome-os-partner:40969
BRANCH=none
TEST=make buildall
CQ-DEPEND=CL:283998

This CL also adds a test example in the extra/ftdi_hostcmd/
directory. To use it, you need the Cr50 attached to the build
host via an FTDI USB-to-SPI adapter.

  cd extra/ftdi_hostcmd
  make
  ./test_cmds

Change-Id: Ia719b1c898afc45b3105a9cd573a8492178d9be2
Signed-off-by: Bill Richardson <wfrichar@chromium.org>