Update patch set 2
Patch Set 2:

At the very least the tests need to be described. There is also need to be a way to easily verify changes in the future. This all is lost now.

Can we revert the changes, discuss the need to change the API, and if necessary - change it, along with tests?

Patch-set: 2