Update patch set 4
Patch Set 4:

Thanks for the review. I'm not in MTV at the moment, therefore I was wondering if someone would be willing to verify the first test for me? It should work (famous last words...), but I just want to be sure. If no one has time, no worries, I'll do it when I return.

Patch-set: 4