Update patch set 2
Patch Set 2: Code-Review-1

as discussed during the original code submition/review, I still think this is not worth the effort to save 0.1% of the buffer space.

On top of that there is a hw bug opened to eliminate access to the mirroring bits in the cr50 SPS controller.

Patch-set: 2
Reviewer: Gerrit User 1000229 <1000229@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=-1