Update patch set 1
Patch Set 1:

Thanks for the comment..Yes I do see that we are enabling and disabling the USB ports as part of power/braswell.c and dont have full knowledge at this point on if these are covered in the SMART power control. I will get back to you on that.

Can we have this merged in and I will get back with removing stuff as another patch?

Patch-set: 1