Update patch set 2
Patch Set 2:

I agree, once the HW is fixed.

It's unlikely, but theoretically possible, that the RX FIFO could fill up completely before the EC can handle it. If that happens, the previous implementation would think it's empty and never do anything about it.

This change adds 32 bytes to the binary size. When the hardware implementation is changed, we can take this back out, and simplify the driver even more by not worrying about the extra bit.

Patch-set: 2