Update patch set 2
Patch Set 2:

"the RX FIFO could fill up completely before the EC can handle it. If that happens, the previous implementation would think it's empty and never do anything about it."

There is no hardware consumer pointer, it is maintained by SW, so this situation is not possible. On top of that, if it were, just just an extra byte would overwrite the first byte in the FIFO and mess up the contents, making it unusable anyways.

IMO this change is not necessary and should be taken out of the stack of patches.

Patch-set: 2