Update patch set 1
Patch Set 1:

You are right. we don't need to have the USB enabled in power/braswell.c. HOOK_CHIPSET_RESUME and HOOK_CHIPSET_SHUTDOWN take care of enabling and disabling the USB in the appropriate power states. 

But HOOK_CHIPSET_RESUME  is happening in S3->S0. and currently we are enabling the USB ports in S5->S3 which means if we remove them USB is powered down during S3.

USB is a wake source in S3 ? I see that its a config option in Rambi CONFIG_USB_PORT_POWER_IN_S3 and this is enabled.

Patch-set: 1