Cr50: Restore Vadim's spstest feature
I accidentally removed this when poking around with the SPS
driver. This adds it back as a separate file. Enabling
CONFIG_SPS_TEST will restore the "spstest" console command to use
for low-level driver tests.

Note that invoking it will replace any other registered SPS
handler.

BUG=chrome-os-partner:40969
BRANCH=none
TEST=manual

Connect the EC to the build host with an FTDI USB-to-SPI adapter.

On the EC console, invoke

  spstest

Build and run Vadim's ftdi_dongle test:

  git clone sso://user/vbendeb/ftdi_dongle
  cd ftdi_dongle/src
  make
  ./examples/spiraw.py -l 10 -f 2000000

Change-Id: Ia6165e3be06d976c59c3e849349da0f7f5006f56
Signed-off-by: Bill Richardson <wfrichar@chromium.org>