Cr50: Fake some data just to test the TPM protocol.
This is an ugly shortcut to test the TPM register-based SPI
protocol. It ignores any incoming data and simply replays a
set of blind responses that cause one specific test to pass.

This is NOT the way to test things thoroughly, but it proves that
the code could at least be made to work someday.

BUG=chrome-os-partner:37753
BRANCH=none
TEST=make buildall

Connect the EC to the build host using an FTDI USB-to-SPI cable,
then run

  sudo trunks_client --ftdi --startup

If nothing has changed recently, that one test should pass.
That's the ONLY test that will pass, though, because we're not
actually doing anything, we're just parroting a set of responses
that we know (knew) will satisfy trunks_client.

Change-Id: I4ded319b55553e4e53a0837e091340b6aff7f569
Signed-off-by: Bill Richardson <wfrichar@chromium.org>