Update patch set 2
Patch Set 2: Code-Review+1

(3 comments)

A few typos.  And I suspect an entertaining discussion when Bill returns.  But since that's not for a week, I won't attempt to channel Bill and hold this up.

It might be easier to review if this were split into 2 changes:

1) A revert of Bill's change

2) The new modifications

(Which might not compile separately from each other, given the new API, but they could CQ-DEPEND on each other so the build system would never know the difference...)

Patch-set: 2
Label: Code-Review=+1