Update patch set 2
Patch Set 2:

(3 comments)

as for doing this int two parts - it is not that easy, there were several changes which went into this file, and they were accompanied by changes to other files and moving sps.h around (not sure why either).

So, this is easier to compare this version file to the known good state to see the actual changes.

The changes are very few and the test is pretty thorough.

Patch-set: 2