Cr50: Enable TPM-protocol data over the SPI bus
The Trusted Computing Group's TPM SPI Hardware Protocol defines a
way to exchange data over a SPI bus by performing reads and
writes to a register space.

This adds a basic implementation of that protocol, which could be
used instead of our EC-style host command.

The only feature implemented at this time is access to the DID_VID
register, which returns vendor ID of 0x1ae0 (Google) and device ID of
0x0028 (cr50).

BUG=chrome-os-partner:37753
BRANCH=none
TEST=make buildall -j succeeds
  - when spstpm is enabled, trunksd properly reads the DID_VID
    register
  - spstest of 100 frames passes

Change-Id: I28d37c3b57dde9adf59e81426efe4f58880cf0b0
Signed-off-by: Bill Richardson <wfrichar@chromium.org>
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>