Discovery: Configure USART2 as a loopback device
This gives a test case for the USART driver on an STM32L.  Eventually
this will be a good place to test that even in a downclocked configuration
the STM32L USART driver can handle 115200 without dropping characters.
This also gives a convenient build test for the STM32L version of the
USART driver.

Signed-off-by: Anton Staaf <robotboy@chromium.org>

BRANCH=None
BUG=None
TEST=make buildall -j
     Cross connect a Discovery and a Discovery-stm32f072

Change-Id: Ifb8dfc1179e8a0be84390d36e0bc3ff15f4f4685