power: ryu: add option to boot on AC plug
Add a new mechanism to optionally boot the main processor when AC power
is plugged on the device.
It is disabled by default, the bootloader might send the
EC_CMD_GSV_BOOT_ON_AC host command to enable/disable it, and the value
survives sysjump.
The whole flow is supposed to work that way :
- the user/lab required value through "fastboot oem off-mode-charge <0|1>"
- depthcharge is storing this value in some non-volatile storage
- on update and EC cold boot (when we wake up the AP to verify a
  possible EC software sync), depthcharge reads this value in its
  non-volatile storage and send it to the EC with EC_CMD_GSV_BOOT_ON_AC
  host command.
- if EC software sync happens, the value is saved by the EC across
  sysjump.

Signed-off-by: Vincent Palatin <vpalatin@chromium.org>

BRANCH=smaug
BUG=chrome-os-partner:41680
TEST=On Smaug EVT2, add some hardcoded EC command in Depthcharge to send
EC_CMD_GSV_BOOT_ON_AC after EC communication initialization and see the
AP booting when plugging a USB cable.

Change-Id: Iecbe0c949ecf27a34ae59df516a1156039b5fa7b