ryu: better external power management
When we are a type-C power sink but we have not decided to charge, cut
the input power path in the charger IC (the USBC_CHARGE_EN_L GPIO was
cutting the charge but still taking power from VBUS for the system).
Update extpower_is_present() to take into account whether we are really
sinking power from our USB-C power source,
and add HOOK_AC_CHANGE events.

The charger chip might reset the Hi-Z bit (disabling the VBUS input
path) once it has done its input detection cycle. In order to mitigate
this, force the current limit to 100mA when in Hi-Z mode and re-verify
the state of the Hi-Z bit when getting an interrupt from the charger.

Signed-off-by: Vincent Palatin <vpalatin@chromium.org>

BRANCH=smaug
BUG=none
TEST=On Ryu EVT2, plug/unplug a dedicated charger, a USB-PD dual role
device and a C-to-A receptacle with a USB key, check everytime in which
direction the current is flowing with Twinkie, the "AC [on|off]"
messages on the console and the battery charging state.

Change-Id: Id90d20a0172c938ef1f15e6a0baf4b5b18bd5ca8