acc: Cleanup SENSOR_INIT_ERROR at shutdown
It used to be that once a sensor got marked with SENSOR_INIT_ERROR it would
never be touched again until the next EC reboot. This would cause intermittent
accelerometer init problems to permanently disable it.

Instead we could forgive the error at shutdown and try again next time the
accelerometers are brought up.

BUG=chrome-os-partner:39269
TEST=Get a flaky accelerometer on minnie,
note how it'll retry initializing it every time we reboot the ap.
BRANCH=veyron

Change-Id: I5e29a939d0992361f14e4c4435cd891857765351
Signed-off-by: Alexandru M Stan <amstan@chromium.org>