Makefile.rules: Fix build timestamp.
The ec_date.h file had incorrect dependencies.  $(objs) had no meaning
outside of the building the object files as it gets privately overidden
with the corresponding target objects (RO, RW, libsharedobjs).  This
caused ec_date.h to only be generated once from a clean.  This commit
fixes that by adding all of the RO and RW objects as dependencies (with
the exception of version.o).

BUG=chrome-os-partner:43373
BRANCH=None
TEST=Built ryu, checked build timestamp in build_info.  Touched a file,
rebuilt, verified that build timestamp was updated.
TEST=make -j buildall tests

Change-Id: I0ab107efc1a504b4f871ebcf595754db1d414c7a
Signed-off-by: Aseda Aboagye <aaboagye@google.com>