TEST-ONLY: glados/glados_pd /board file changes required to test VBUS ALERT
Detecting VBUS via the TCPM isn't required for Glados as it has GPIO signals
connected to the EC MCU. However, in order to test that communicating VBUS
status via the TCPM/TCPCI interface works, or to use this feature, these
changes would be required.

BUG=none
BRANCH=none
TEST=Manual test with Zinger to Glados and Samus to Glados connections.
Verified that VBUS changes were detected on the PD MCU and then signaled to the
EC MCU via the ALERT# signal and that the TCPM communicated the correct
VBUS value to the PD layer.

Change-Id: I44f5eeff24c5fff044a67523cffe6ace0da7f027
Signed-off-by: Scott Collyer <scollyer@chromium.org>